test: (#829) MemberControllerTest -> AuthControllerTest로 컨트롤러 단위테스트 이동
