+++ backend/src/test/java/com/votogether/domain/member/controller/MemberControllerTest.java
@@ -240,24 +240,4 @@ class MemberControllerTest extends ControllerTest {
                 .statusCode(HttpStatus.OK.value());
     }
 
-    @Test
-    @DisplayName("회원 탈퇴에 성공하면 204를 반환한다.")
-    void deleteMember() throws Exception {
-        // given
-        TokenPayload tokenPayload = new TokenPayload(1L, 1L, 1L);
-
-        given(tokenProcessor.resolveToken(anyString())).willReturn("token");
-        given(tokenProcessor.parseToken(anyString())).willReturn(tokenPayload);
-        given(memberService.findById(anyLong())).willReturn(MemberFixtures.FEMALE_20.get());
-        willDoNothing().given(memberService).deleteMember(MemberFixtures.FEMALE_20.get());
-
-        // when, then
-        RestAssuredMockMvc
-                .given().log().all()
-                .headers(HttpHeaders.AUTHORIZATION, "Bearer token")
-                .when().delete("/members/me/delete")
-                .then().log().all()
-                .statusCode(HttpStatus.NO_CONTENT.value());
-    }
-
 }
